+++ dev/Neuron.js
@@ -11,6 +11,7 @@ class Neuron {
     }
 
     init (size, {adaptiveLR, activationConfig, eluAlpha}={}) {
+        
         if(!this.imported){
             this.weights = [...new Array(size)].map(v => Math.random()*0.2-0.1)
             this.bias = Math.random()*0.2-0.1