+++ CHANGELOG.md
@@ -1,15 +1,17 @@
-# Upcoming - Activation functions
+# v1.2.0 - Activation functions (Part 1)
 ---
+#### Network
+- Added lreluSlope, for lrelu activation
+- Added eluAlpha, for elu activation
 #### NetMath
 - Added tanh activation function
 - Added relu activation function
 - Added lrelu activation function
-- Added lreluSlope as a network configuration
 - Added rrelu activation function
 - Added lecuntanh activation function
 - Added sech to NetMath
 - Added elu activation function
-- Added eluAlpha as a network configuration, for elu
+
 # v1.1.0 - Adaptive learning rates
 ---
 #### Network